refactor: move useTab to its own folder.
