+++ packages/mui-base/src/TabUnstyled/index.ts
@@ -3,6 +3,3 @@ export * from './TabUnstyled.types';
 
 export { default as tabUnstyledClasses } from './tabUnstyledClasses';
 export * from './tabUnstyledClasses';
-
-export { default as useTab } from './useTab';
-export * from './useTab.types';
