+++ packages/mui-base/src/index.js
@@ -26,6 +26,8 @@ export * from './useTabsList';
 
 export * from './useTabs';
 
+export * from './useTab';
+
 export { default as BadgeUnstyled } from './BadgeUnstyled';
 export * from './BadgeUnstyled';
