+++ packages/mui-base/src/useTab/index.ts
@@ -1,0 +1,5 @@
+export { default as useTab } from './useTab';
+export { default } from './useTab';
+export * from './useTab';
+
+export * from './useTab.types';
