+++ packages/mui-joy/src/Tab/Tab.tsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { OverridableComponent } from '@mui/types';
 import { unstable_capitalize as capitalize, unstable_useForkRef as useForkRef } from '@mui/utils';
 import { unstable_composeClasses as composeClasses } from '@mui/base';
-import { useTab } from '@mui/base/TabUnstyled';
+import useTab from '@mui/base/useTab';
 import { useSlotProps } from '@mui/base/utils';
 import { ListItemButtonRoot } from '../ListItemButton/ListItemButton';
 import { useThemeProps } from '../styles';
